requirements.txt
tensorflow>=2.14
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
opencv-python>=4.7
tensorflow-addons>=0.21
